+++ src/codemagic/tools/android_keystore/android_keystore.py
@@ -1,12 +1,15 @@
 from __future__ import annotations
 
+import json
 import pathlib
+from typing import List
 from typing import Optional
 from typing import Union
 
 from codemagic import cli
 from codemagic.cli import Colors
 from codemagic.mixins import PathFinderMixin
+from codemagic.models import Certificate
 from codemagic.models import CertificateAttributes
 from codemagic.models import Keystore
 from codemagic.shell_tools import Keytool
@@ -29,14 +32,14 @@ class AndroidKeystore(cli.CliApp, PathFinderMixin):
 
     @classmethod
     def _get_certificate_attributes(
-            cls,
-            issuer_common_name: Optional[str],
-            issuer_organization: Optional[str],
-            issuer_organization_unit: Optional[str],
-            issuer_locality: Optional[str],
-            issuer_state_or_province: Optional[str],
-            issuer_country: Optional[str],
-            issuer_distinguished_name: Optional[str],
+        cls,
+        issuer_common_name: Optional[str],
+        issuer_organization: Optional[str],
+        issuer_organization_unit: Optional[str],
+        issuer_locality: Optional[str],
+        issuer_state_or_province: Optional[str],
+        issuer_country: Optional[str],
+        issuer_distinguished_name: Optional[str],
     ) -> CertificateAttributes:
         if issuer_distinguished_name is None:
             certificate_attributes = CertificateAttributes(
@@ -75,20 +78,20 @@ class AndroidKeystore(cli.CliApp, PathFinderMixin):
         ),
     )
     def create(
-            self,
-            keystore_path: pathlib.Path,
-            keystore_password: Union[str, KeystorePassword],
-            key_alias: str,
-            key_password: Optional[Union[str, KeyPassword]] = None,
-            issuer_common_name: Optional[str] = None,
-            issuer_organization: Optional[str] = None,
-            issuer_organization_unit: Optional[str] = None,
-            issuer_locality: Optional[str] = None,
-            issuer_state_or_province: Optional[str] = None,
-            issuer_country: Optional[str] = None,
-            issuer_distinguished_name: Optional[str] = None,
-            overwrite_existing: bool = False,
-            validity_days: int = CreateAndroidKeystoreArgument.VALIDITY_DAYS.get_default(),
+        self,
+        keystore_path: pathlib.Path,
+        keystore_password: Union[str, KeystorePassword],
+        key_alias: str,
+        key_password: Optional[Union[str, KeyPassword]] = None,
+        issuer_common_name: Optional[str] = None,
+        issuer_organization: Optional[str] = None,
+        issuer_organization_unit: Optional[str] = None,
+        issuer_locality: Optional[str] = None,
+        issuer_state_or_province: Optional[str] = None,
+        issuer_country: Optional[str] = None,
+        issuer_distinguished_name: Optional[str] = None,
+        overwrite_existing: bool = False,
+        validity_days: int = CreateAndroidKeystoreArgument.VALIDITY_DAYS.get_default(),
     ) -> Keystore:
         """
         Create an Android keystore
@@ -139,9 +142,9 @@ class AndroidKeystore(cli.CliApp, PathFinderMixin):
         CreateAndroidKeystoreArgument.VALIDITY_DAYS,
     )
     def create_debug_keystore(
-            self,
-            overwrite_existing: bool = False,
-            validity_days: int = CreateAndroidKeystoreArgument.VALIDITY_DAYS.get_default(),
+        self,
+        overwrite_existing: bool = False,
+        validity_days: int = CreateAndroidKeystoreArgument.VALIDITY_DAYS.get_default(),
     ) -> Keystore:
         """
         Create Android debug keystore at ~/.android/debug.keystore
@@ -162,45 +165,85 @@ class AndroidKeystore(cli.CliApp, PathFinderMixin):
         AndroidKeystoreArgument.KEYSTORE_PATH,
         AndroidKeystoreArgument.KEYSTORE_PASSWORD,
         AndroidKeystoreArgument.KEY_ALIAS,
-        AndroidKeystoreArgument.KEY_PASSWORD,
     )
     def verify(
-            self,
-            keystore_path: pathlib.Path,
-            keystore_password: Union[str, KeystorePassword],
-            key_alias: str,
-            key_password: Optional[Union[str, KeyPassword]] = None,
+        self,
+        keystore_path: pathlib.Path,
+        keystore_password: Union[str, KeystorePassword],
+        key_alias: str,
     ):
         """
         Check that the keystore password, key password and key alias are correct
         for specified keystore
         """
         store_password: str = KeystorePassword.resolve_value(keystore_password)
-        if key_password is None:
-            _key_password: str = store_password
-        else:
-            _key_password = KeyPassword.resolve_value(key_password)
 
-        keystore = Keystore(
-            key_alias=key_alias,
-            key_password=_key_password,
-            store_path=keystore_path,
-            store_password=store_password,
-        )
-
-        self.logger.info(f'Verify Android keystore at "{keystore.store_path}"')
-        if not keystore.store_path.is_file():
-            error = f'Keystore does not exists at {keystore.store_path}'
-            raise AndroidKeystoreArgument.KEYSTORE_PATH.raise_argument_error(error)
+        self.logger.info(f'Verify Android keystore at "{keystore_path}"')
+        self._assert_keystore_exists(keystore_path)
 
         try:
-            Keytool().validate_keystore(keystore)
+            Keytool().validate_keystore(
+                keystore_path=keystore_path,
+                keystore_password=store_password,
+                key_alias=key_alias,
+            )
         except ValueError as ve:
             raise AndroidKeystoreError(str(ve)) from ve
         self.logger.info((
-            f'Keystore "{keystore.store_path}" has alias "{keystore.key_alias}" '
-            f'and can be unlocked with given passwords'
+            f'Keystore "{keystore_path}" has alias "{key_alias}" '
+            f'and can be unlocked with given password'
         ))
+
+    @cli.action(
+        'certificates',
+        AndroidKeystoreArgument.KEYSTORE_PATH,
+        AndroidKeystoreArgument.KEYSTORE_PASSWORD,
+        AndroidKeystoreArgument.KEY_ALIAS_OPTIONAL,
+        AndroidKeystoreArgument.JSON_OUTPUT,
+    )
+    def list_certificates(
+        self,
+        keystore_path: pathlib.Path,
+        keystore_password: Union[str, KeystorePassword],
+        key_alias: Optional[str] = None,
+        json_output: bool = False,
+        should_print: bool = True,
+    ) -> List[Certificate]:
+        """
+        List certificates that are included in keystore
+        """
+        store_password: str = KeystorePassword.resolve_value(keystore_password)
+
+        if key_alias is None:
+            self.logger.info(f'List certificates in Android keystore "{keystore_path}"')
+        else:
+            self.logger.info(f'List certificates in Android keystore "{keystore_path}" for alias "{key_alias}"')
+
+        self._assert_keystore_exists(keystore_path)
+
+        try:
+            certificates = Keytool().get_certificates(
+                keystore_path=keystore_path,
+                keystore_password=store_password,
+                key_alias=key_alias,
+            )
+        except ValueError as ve:
+            raise AndroidKeystoreError(str(ve)) from ve
+
+        if should_print:
+            if json_output:
+                summary = json.dumps([c.get_summary() for c in certificates], indent=4)
+            else:
+                summary = '\n'.join([c.get_text_summary() for c in certificates])
+            self.echo(summary)
+
+        return certificates
+
+    @classmethod
+    def _assert_keystore_exists(cls, keystore_path: pathlib.Path):
+        if not keystore_path.is_file():
+            error = f'Keystore does not exists at {keystore_path}'
+            raise AndroidKeystoreArgument.KEYSTORE_PATH.raise_argument_error(error)
 
 
 if __name__ == '__main__':